+++ src/main/java/com/tensai/cms/telegram/api/events/UpdateMessageEvent.java
@@ -14,7 +14,7 @@ public record UpdateMessageEvent(
         Long chatId,
 
         @JsonProperty("message_thread_id")
-        Integer messageThreadId,
+        Long messageThreadId,
 
         @JsonProperty("message_id")
         Integer messageId,
